requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
python-dotenv==1.0.0
matplotlib==3.8.2
plotly==5.18.0
pydantic>=2.7.2,<3.0.0
sentence-transformers==2.6.1
numpy==1.26.4
uvicorn==0.29.0
fastapi==0.110.2
fastapi-mcp  # make sure it's a version that works with MCP >= 1.4.1
mcp>=1.4.1   # requires pydantic >=2.7.2
google-adk
